Wire-format widening: origin_hash u32 → u64 across application layer
Closes the audit's `origin_hash: u32` collision-floor item. The
prior u32 truncation produced a ~65k peer birthday-collision
floor in cross-channel accounting and chain-integrity checks
that aliased distinct entities. Widening to u64 lifts the floor
to ~2⁶⁴.

`EntityId::origin_hash` now derives 8 bytes from the same
BLAKE2s-MAC (`b"net-origin-v1"` domain label unchanged), so the
low 32 bits are byte-identical to the prior u32 derivation —
existing `NetHeader` consumers see the same routing bytes.

Wire-format bumps:
  - CAUSAL_LINK_SIZE 28 → 32
  - EVENT_META_SIZE 20 → 24
  - CONTINUITY_PROOF_SIZE 36 → 40
  - ForkRecord wire prefix 25 → 33

Per-packet `NetHeader::origin_hash` STAYS u32. It's a routing
fast-path field; collisions there are benign (the receiver's
deeper chain checks would catch any aliasing) and bumping per-
packet header has a much larger blast radius. Callers downcast
`kp.origin_hash() as u32` at the `NetHeader` boundary; the
matched low-32-bits derivation means the bytes-on-wire are
identical to before.

Application-layer types widened to u64 throughout:
  - EntityId, EntityKeypair, OriginStamp accessors + cached fields
  - CausalLink struct field + wire encoders/decoders
  - EventMeta struct field + wire encoders/decoders +
    for_checksum_bytes (checksum slot moved from [16..20] to [20..24])
  - CausalChainBuilder, ChainError::OriginMismatch fields
  - All migration-message wire encoders/decoders for
    daemon_origin (size-precondition checks bumped 4→8 bytes per occurrence)
  - DashMap<u32, _> / HashMap<u32, _> / HashSet<u32> /
    Vec<(usize, u32)> keyed by origin/daemon → u64 generic
  - DaemonError::NotFound / Stale variant fields → u64
  - MigrationError::DaemonNotFound / AlreadyMigrating → u64
  - EntityLog::origin_hash, LogIndex Ref types
  - CausalCone, SuperpositionState, ForkRecord fields + accessors
  - GroupCoord::origin_hash_for_entity_id + route_event return
  - ReplicaGroup::route_event return + ForkGroup parent_origin
  - Migration handler callback type aliases (PreCleanup,
    Failure, PostRestore)
  - SnapshotReassembler key tuple (u32, u64) → (u64, u64)
  - cortex/memories + cortex/tasks adapter origin_hash field

FFI:
  - net_identity_origin_hash returns uint64_t (header
    net.go.h updated to match)
  - cortex FFI signatures updated for adapter.open(origin_hash: u64)

No backwards compat — pre-bump bytes (snapshots, persisted
RedEX events, in-flight migration messages) cannot be read by
post-bump binaries. The migration cliff is documented in this
commit message; the BLAKE2s domain label is intentionally
unchanged so the routing-low-32 bits stay identical for any
peer running both old and new code via NetHeader (only the
new application-layer types differ).

Build status: 1391 lib tests pass, full integration suite
passes (~1605 tests total), `cargo clippy --all-features
--all-targets -- -D warnings` clean.

Audit final state: every actionable item is fixed, obsolete, or
skipped with documented reasoning. Remaining-Critical: 0.
Remaining-High: 0. Remaining-Medium/Lower: 0.

+++ net/crates/net/benches/cortex.rs
@@ -21,7 +21,7 @@ use net::adapter::net::netdb::NetDbSnapshot;
 use net::adapter::net::redex::Redex;
 use tokio::runtime::Runtime;
 
-const ORIGIN: u32 = 0xABCD_EF01;
+const ORIGIN: u64 = 0xABCD_EF01;
 
 /// A shared tokio runtime for all bench scenarios. `CortexAdapter::open`
 /// and every ingest path touch `tokio::spawn` / `broadcast::Sender::send`
